Fixed reverse non-functionality.
Requires jumpering fan speed pin to A+, and encoder input B1 to B+. I acheived this with a piece of 22 ga. wire and a pin crimp fitting.

Not an ideal solution, but trying to make a board to go between the ATMega328 and its socket to switch the pins around didn't work, and I was out of parts with which to try again. I'll take another crack at it when I next order a set of PCBs.

Now, enough technical details. It's story time....

A lot of pain went into this commit.

In testing this, I accidentally jumpered the two pins in question to either side of channel A. This resulted in the robot booting, and promptly pinning the throttle in full reverse for the duration of the boot chirp sequence. Once I had cleaned the skidmarks off the floor, returned my pulse to a regular rate, and fixed the lamp that it had careened into, I went back and double checked the code. I also set a replacement bearing block printing, since the rear-left half-axle had taken the majority of the blow, and suffered dearly for it.

Everything seemed alright, so I took the chain off of right side (channel B) and disconnected the positive lead to the left side (channel A). I then propped the vehicle up on a box and braced the front end on my left leg.

What I failed to notice was that the stakeon I had disconnected from motor B had opted to rub up against the exposed source/tab of the channel-A  H-bridge. Which is what it's supposed to connect to in the first place. So, with a bright flash, a loud noise, and several more skidmarks, the robot had crashed into the wall again, with only one working motor. In the process, it gave me the mother of all friction burns on my left knee and right ankle, both of which it ran over with the benefit of 500W of full throttle.

While cleaning up the skidmarks - and sanitizing the wounds - from that experience, I noticed that I had put the jumpers in the wrong spots. After a moment of introspection ("Charles," I pondered, "why are you such a melonhead at times?") I swapped the pins, and geared up for another test run. This time, however, I took off both chains before powering anything on, and did not place any parts of my body between it and the nearest wall.

This time, with all the precautions in place for a serious failure.... everything worked. So, breathing a sigh of relief, I went back in and changed the logic for controlling reverse motion, uploaded it, and gave it a go.

Victory, and at last an end to the bloodshed.

Totally worth it.